fix: script rename
