+++ src/client/SentenceLog.ts
